feat(lambdas): :sparkles: update import file parser to send csv rows to sqs
